Se modifico el juego para agregar aleatoriodad al alien

+++ game/alien.py
@@ -3,7 +3,7 @@ import pygame
 
 
 class Alien(pygame.sprite.Sprite):
-    def __init__(self, x: int, y: int, type: int):
+    def __init__(self, x: float, y: int, type: int):
         super().__init__()
         # El tipo carga el tipo de alien
         self.type = type
